pyproject.toml
[tool.poetry]
name = "webapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
flask = "^3.0.3"
torch = "^2.4.1"
torchvision = "^0.19.1"
pillow = "^10.4.0"
streamlit = "^1.38.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
